+++ transport-native-io_uring/src/main/java/io/netty/channel/uring/IOUring.java
@@ -28,7 +28,7 @@ final class IOUring {
     static final int OP_WRITE = 23;
     static final int OP_POLL_REMOVE = 7;
 
-    static final int POLLMASK_LINK = 1;
+    static final int POLLMASK_IN = 1;
     static final int POLLMASK_OUT = 4;
     static final int POLLMASK_RDHUP = 8192;
